+++ _pytest/_code/code.py
@@ -140,8 +140,8 @@ class TracebackEntry(object):
     _repr_style = None
     exprinfo = None
 
-    def __init__(self, rawentry, exctype=None):
-        self._exctype = exctype
+    def __init__(self, rawentry, excinfo=None):
+        self._excinfo = excinfo
         self._rawentry = rawentry
         self.lineno = rawentry.tb_lineno - 1
 
@@ -218,16 +218,12 @@ class TracebackEntry(object):
 
     source = property(getsource)
 
-    def _is_exception_type(self, obj):
-        return isinstance(obj, type) and issubclass(obj, Exception)
-
     def ishidden(self):
         """ return True if the current frame has a var __tracebackhide__
             resolving to True
 
-            If __tracebackhide__ is set to an exception type, or a list/tuple,
-            the traceback is only hidden if the exception which happened is of
-            the given type(s).
+            If __tracebackhide__ is a callable, it gets called with the
+            ExceptionInfo instance and can decide whether to hide the traceback.
 
             mostly for internal use
         """
@@ -239,13 +235,8 @@ class TracebackEntry(object):
             except KeyError:
                 return False
 
-        if self._is_exception_type(tbh):
-            assert self._exctype is not None
-            return issubclass(self._exctype, tbh)
-        elif (isinstance(tbh, (list, tuple)) and
-              all(self._is_exception_type(e) for e in tbh)):
-            assert self._exctype is not None
-            return issubclass(self._exctype, tuple(tbh))
+        if callable(tbh):
+            return tbh(self._excinfo)
         else:
             return tbh
 
@@ -272,13 +263,13 @@ class Traceback(list):
         access to Traceback entries.
     """
     Entry = TracebackEntry
-    def __init__(self, tb, exctype=None):
-        """ initialize from given python traceback object and exc type. """
-        self._exctype = exctype
+    def __init__(self, tb, excinfo=None):
+        """ initialize from given python traceback object and ExceptionInfo """
+        self._excinfo = excinfo
         if hasattr(tb, 'tb_next'):
             def f(cur):
                 while cur is not None:
-                    yield self.Entry(cur, exctype=exctype)
+                    yield self.Entry(cur, excinfo=excinfo)
                     cur = cur.tb_next
             list.__init__(self, f(tb))
         else:
@@ -302,7 +293,7 @@ class Traceback(list):
                  not codepath.relto(excludepath)) and
                 (lineno is None or x.lineno == lineno) and
                 (firstlineno is None or x.frame.code.firstlineno == firstlineno)):
-                return Traceback(x._rawentry, self._exctype)
+                return Traceback(x._rawentry, self._excinfo)
         return self
 
     def __getitem__(self, key):
@@ -321,7 +312,7 @@ class Traceback(list):
             by default this removes all the TracebackItems which are hidden
             (see ishidden() above)
         """
-        return Traceback(filter(fn, self), self._exctype)
+        return Traceback(filter(fn, self), self._excinfo)
 
     def getcrashentry(self):
         """ return last non-hidden traceback entry that lead
@@ -385,7 +376,7 @@ class ExceptionInfo(object):
         #: the exception type name
         self.typename = self.type.__name__
         #: the exception traceback (_pytest._code.Traceback instance)
-        self.traceback = _pytest._code.Traceback(self.tb, exctype=self.type)
+        self.traceback = _pytest._code.Traceback(self.tb, excinfo=self)
 
     def __repr__(self):
         return "<ExceptionInfo %s tblen=%d>" % (self.typename, len(self.traceback))