+++ CHANGELOG.rst
@@ -23,8 +23,9 @@
   Thanks `@omarkohl`_ for the complete PR (`#1502`_) and `@nicoddemus`_ for the
   implementation tips.
 
-* ``__tracebackhide__`` can now also be set to an exception type (or a list of
-  exception types) to only filter exceptions of that type.
+* ``__tracebackhide__`` can now also be set to a callable which then can decide
+  whether to filter the traceback based on the ``ExceptionInfo`` object passed
+  to it.
 
 *